Move things from voice to channel
